% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\docType{package}
\name{ppgam-package}
\alias{ppgam-package}
\title{ppgam: A package for fitting generalised additive point process models.}
\description{
The ppgam package's main function is ppgam.
}
\section{ppgam functions}{

The ppgam functions ...
}

\author{
\strong{Maintainer}: Ben Youngman \email{b.youngman@exeter.ac.uk}

Authors:
\itemize{
  \item Ben Youngman \email{b.youngman@exeter.ac.uk}
  \item Theo Economou
}

}
\keyword{internal}
