% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USlandfall}
\alias{USlandfall}
\title{Times of landfalling US hurricanes}
\format{
A data frame with 61129 rows and 2 variables

The variables are as follows:

\describe{
  \item{date}{date of landfall, as class "Date"}
  \item{landfall}{an integer: did a hurricane make landfall on this day?}
}
}
\usage{
data(USlandfall)
}
\description{
A data frame:
}
\examples{

data(USlandfall)
plot(USlandfall, type="h")

}
\references{
https://www.nhc.noaa.gov/data/
}
\keyword{datasets}
